Moved 'add trips' files to ViewTrips folder and updated imports.
