+++ frontend/src/components/ViewTrips/add-trip.js
