+++ frontend/src/components/ViewTrips/create-new-trip.js
